### 2024-06-13 #154 @editorialbot
[PRE REVIEW]: BlankLocalizationCore.jl: implementing blank localization in Julia
**Submitting author:** @cserteGT3 (<a href="http://orcid.org/0000-0002-8028-8061">Tamás Cserteg</a>)
**Repository:** https://github.com/cserteGT3/BlankLocalizationCore.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.3.0
**Managing EiC:** Mathieu Besançon

Thanks for submitting your paper to JCON @cserteGT3. **Currently, there isn't a JCON editor assigned** to your paper.
