### 2024-07-02 a7ee393
CompatHelper: bump compat for Meshes to 0.45, (keep existing compat)
