+++ Project.toml
@@ -16,15 +16,15 @@ Rotations = "6038ab10-8711-5258-84ad-4b1120ba62dc"
 [compat]
 DataFrames = "1"
 JuMP = "1.4"
-Meshes = "0.35,0.36"
+Meshes = "0.35,0.36, 0.45"
 PrettyTables = "2"
 Rotations = "1.5"
 julia = "1.9"
 
 [extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 Meshes = "eacbb407-ea5a-433e-ab97-5258b1ca43fa"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test", "Ipopt", "Meshes"]
