### 2024-08-02 63aebc6
CompatHelper: bump compat for Meshes to 0.47, (keep existing compat)
