### 2024-09-02 4faddb0
CompatHelper: bump compat for Meshes to 0.50, (keep existing compat)
