### 2024-11-02 08e5e41
CompatHelper: bump compat for Meshes to 0.52, (keep existing compat)
