### 2025-02-26 #154 @mzagorowska
> Hi [@mzagorowska](https://github.com/mzagorowska)👋,
>
> would you like to review this JuliaCon proceedings submission?
>
> you can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).

Sure
