### 2025-02-27 #154 @Nimrais
@fcdimitr Sure, I'd be happy to review this JuliaCon proceedings submission!
