### 2025-02-27 #182 @editorialbot
[REVIEW]: BlankLocalizationCore.jl: implementing blank localization in Julia
**Submitting author:** @cserteGT3 (<a href="http://orcid.org/0000-0002-8028-8061">Tamás Cserteg</a>)
**Repository:** https://github.com/cserteGT3/BlankLocalizationCore.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.4.1
