### 2025-03-04 #182 @mzagorowska
General comments regarding the missing ticks:
- (installation, installation instructions) I had to add packages (ipopt, GLMakie, Meshes) from errors when trying to run the example
- (performance) It is not entirely clear to me whether the improvements shown in Fig 2 are from the algorithm published already in Ref. [4], or from the implementation in Julia in the current work. Also, the plots in the current paper seem to be taken 1:1 from Ref [4]
- (methodology) The approach is justified, the trouble is with reproducibility based purely on the paper.
1. The paper refers to Ref [4] for the algorithm and I would find it difficult to say what exactly is done without Ref [4].
2. I would also like to see a bit more of detail on the interplay between the packages and the custom extensions that made using diverse packages, such as JuMP with Meshes, possible.
3. It could also be a good addition to say how the elements from Section 2 were implemented in Julia (at least mention the relevant functions). Right now, the first part is a rather theoretical introduction to the problem, without indicating the solutions in Julia.
4. Regarding future plans, it would also be interesting to see how the proposed extensions can be included in the package, if they need substantial changes.
