### 2025-03-24 d4d7eed
fix visualization issue in example
