### 2025-03-24 1bff2d3
fix visualization issue in example
