### 2025-03-24 16496ed
fix installation instructions
