+++ docs/src/example.md
@@ -1,12 +1,23 @@
 # Example
 
-This example can be found as a single file in the repository under the `examples` folder.
-The only difference is, that the file uses the [Ipopt](https://github.com/jump-dev/Ipopt.jl) package for solving the optimization.
+This example can be found as a single file in the repository under the [`examples`](https://github.com/cserteGT3/BlankLocalizationCore.jl/blob/main/examples/example.jl) folder.
+The script file does not contain the package installation steps, though it comes with a `Project.toml` file that you can instantiate.
+
+## Getting started
+
+Install the necessary packages:
+
+* `Ipopt` (or `Xpress`)
+* `Meshes` and `GLMakie`: only necessary for visualization
+
+```julia
+] add BlankLocalizationCore Ipopt Meshes GLMakie
+```
 
 ## Example part
 
 The following part is crafted for demonstrating purposes.
-The CAD filesa are available here: [machined](assets/example-part-machined.stl) and [rough](assets/example-part-rough.stl).
+The CAD files are available here: [machined](assets/example-part-machined.stl) and [rough](assets/example-part-rough.stl).
 The two images below show the part from its "front" and "back" sides in its machined state.
 
 ![Machined look front](assets/machined-front.png)
@@ -263,11 +274,10 @@ These are:
 | SetPartZeroPosition | The position of each part zero can be set with this option. A vector of 3 long vectors is expected, that is matched with the number of part zeros. Empty vectors can be passed, if not all part zero positions should be set. For example setting this option to `[[], [], []]` for three part zeros will have no effect. `NaN` elements are ignored, which means that specific axes of part zeros can be set. For example `[[NaN, 150, Nan], [], []]` would only set the value of the Y axis of the first part zero, all others are unaffected. | | Optional |
 | maxPlaneZAllowance | The allowance has no upper bound by default. For planes, it is possible to set an upper bound to avoid the issue of "machining the whole part away". | | Optional |
 
-We need to load the JuMP package and also an optimizer.
-For the papers we used FICO Xpress, therefore here it will be used as well, thanks to the [Xpress.jl](https://github.com/jump-dev/Xpress.jl) package.
-Note that a licence is needed, however the generated problem size does not exceed the limit of the (free) community license.
-
+An optimization solver is needed for which we use [Xpress.jl](https://github.com/jump-dev/Xpress.jl) in our research.
+A license is required for FICO Xpress (though the generated problem size does not exceed the limit of the (free) community license).
 The [Ipopt](https://github.com/jump-dev/Ipopt.jl) and [SCIP](https://github.com/scipopt/SCIP.jl) solvers were also tested and can be used.
+We use the former in this example.
 
 ```julia
 ## Constructing and solving the optimization problem
@@ -276,16 +286,16 @@ pard = Dict("minAllowance"=>0.5, "OptimizeForToleranceCenter"=>true,
 
 mop = MultiOperationProblem(partzeros, holes, planes, tolerances, pard)
 
-import JuMP
-import Xpress
+import Ipopt # or Xpress if you have a license
+optimizeproblem!(mop, Ipopt.Optimizer)
 
+#=
[…]
