### 2025-03-24 6188c5f
fix installation instructions
