### 2025-03-25 1c4390b
update CI rules run test job on all tags, branches and docs job only on main branch and tags
