### 2025-03-25 97aa435
be more permissive on job run, as Documenter will handle deployment
