+++ paper/paper.tex
@@ -6,6 +6,7 @@
 \begin{document}
 
 \input{header}
+%TODO: ha kell hely, akkor Jóska BME-t törölhetjük
 
 \maketitle
 
@@ -13,7 +14,9 @@
 
 Blank localization (also known as workpiece referencing) is an essential task in machining.
 It aims to precisely establish the geometric relation of the machine tool (mill, lathe, etc.) and the workpiece.
-We introduced the concept of multi-operation blank localization to address this task for drilling and milling scenarios in a semi-automated way, which allows positioning different machining features (e.g., different holes) separately in order to exploit the tolerances on the relative position of those features to compensate the small errors of the blank.
+We introduced the concept of multi-operation blank localization to address this task for drilling and milling scenarios in a semi-automated way,
+%TODO: innentől a mondat második fele cut?
+which allows positioning different machining features (e.g., different holes) separately in order to exploit the tolerances on the relative position of those features to compensate the small errors of the blank.
 The method takes as input the measured rough geometry and the machining CNC code, and computes the best possible position of each feature considering machining allowances and tolerances by solving a convex quadratically constrained quadratic program (QCQP).
 The versatility and extensibility of the Julia language helped the development of this algorithm, materializing in the \texttt{BlankLocalizationCore.jl} package.
 Its flexibility and ease of use make it an excellent research tool that can be deployed in production as well.
@@ -23,29 +26,44 @@ Its flexibility and ease of use make it an excellent research tool that can be d
 \label{sec:intro}
 Cast parts may have small geometric variations from lot to lot that need to be addressed before machining by altering the CNC code.
 Current practice is dominated by iterative adjustments by the human operator, which requires highly trained workers, takes a long time, and still, may produce scrap.
-Automated methods exist for complex free-form parts like wind turbines that place the entire blank as a single solid object (\cite{ding:2021_CoarsefineOptimizationMethod}\cite{tan:2014_UnconstrainedApproachBlank}).
-Multi-operation blank localization was introduced in~\cite{cserteg:2023_Annals} focusing on drilling and milling.
-These machining operations are among the most used ones, therefore the method is applicable for a wide range of products.
-The abstract method and its implementation was developed in parallel, which required a language with support for easy prototyping and wide variety of tools.
+%TODO: kell egyáltalán ezeket hivatkozni?
+Automated methods exist for complex free-form parts like wind turbines that place the entire blank as a single solid object \cite{ding:2021_CoarsefineOptimizationMethod}\cite{tan:2014_UnconstrainedApproachBlank}.
+Multi-operation blank localization~\cite{cserteg:2023_Annals} however handles group of features independently providing greater flexibility than traditional approaches.
+It focuses on drilling and milling which are among the most common machining operations, making it applicable to a wide range of products.
+The abstract method and its implementation were developed in parallel, which required a language with support for easy prototyping and wide variety of tools.
 Exactly for these reasons we chose the Julia language~\cite{bezanson2017julia}.
 
 \section{Multi-operation blank localization}
 \label{sec:algo}
 
-The algorithm looks for the optimal position for each feature group on the workpiece.
-Features within a group can be, for example, drilled holes on the same side of the part, whose drilling positions are defined relative to a common reference (the part zero).
-Features within the same group move together, but different feature groups can be moved separately.
-The to-be-machined features (short: machined features) need to "enclose" the features on the blank (called rough features) with a minimum machining allowance which ensures the required surface finish.
[…]
