+++ paper/paper.tex
@@ -25,42 +25,67 @@ Its flexibility and ease of use make it an excellent research tool that can be d
 \section{Introduction}
 \label{sec:intro}
 Cast parts may have small geometric variations from lot to lot that need to be addressed before machining by altering the CNC code.
-Current practice is dominated by iterative adjustments by the human operator, which requires highly trained workers, takes a long time, and still, may produce scrap.
+Current practice is dominated by iterative adjustments by human operators, which requires highly trained workers and takes a long time.
+%, and still, may produce scrap.
 %TODO: kell egyáltalán ezeket hivatkozni?
 Automated methods exist for complex free-form parts like wind turbines that place the entire blank as a single solid object \cite{ding:2021_CoarsefineOptimizationMethod}\cite{tan:2014_UnconstrainedApproachBlank}.
 Multi-operation blank localization~\cite{cserteg:2023_Annals} however handles group of features independently providing greater flexibility than traditional approaches.
 It focuses on drilling and milling which are among the most common machining operations, making it applicable to a wide range of products.
-The abstract method and its implementation were developed in parallel, which required a language with support for easy prototyping and wide variety of tools.
+The abstract method and its implementation were developed in parallel, which required a language with wide variety of tools and support for easy prototyping.
 Exactly for these reasons we chose the Julia language~\cite{bezanson2017julia}.
 
 \section{Multi-operation blank localization}
 \label{sec:algo}
 
 The problem involves looking for the optimal position for each to-be-machined feature (short: machined feature) on the workpiece.
-Based on the actual machining operations, these features are grouped into feature groups, for example, drilled holes on the same side of the part, whose drilling positions are defined relative to a common reference (the part zero).
-Features within the same group move together, but different feature groups can move separately.
-The machined features need to "enclose" the pre-cast features on the blank (called rough features) with a minimum machining allowance which ensures the required surface finish.
-While doing so, the dimensional tolerances between some features need to be respected, as these describe functional properties (e.g. connections to mating parts).
-Fig.~\ref{fig:hatfig} from \cite{cserteg:2023_Annals} shows an example of these features.
+%chatgpt
+These features are grouped together, with each group defined relative to a specific reference point known as the part zero.
+The layout of these part zeros is controlled by the structure of the machining CNC code.
+By moving the part zeros, we can indirectly control the positions of the associated features.
+Each part zero—and by extension, each group—can be moved independently, which gives the method its flexibility.
+In the optimization program, the positions of the part zeros are the decision variables.
 
-\begin{figure}[hbt]
+%eredeti
+\iffalse
+These features are structured into groups, each group being defined relative to a common reference, called part zero.
+This is defined by the structure of the machining CNC code.
+This means that we can influence the features' positions through moving the part zeros.
+Part zeros (thus the groups) can be moved separately, which gives the method its flexibility.
+The position of the part zeros are the decision variables of the optimization program.
+\fi
+%Features within the same group must be moved together, but different feature groups can be moved separately.
+%TODO: optimalizáló leírás:
+% 0. motiváció
+% 1. mik a döntések
+% 2. mik a korlátok
+% 3. mi a célfv
+
+The machined features need to "enclose" the pre-cast features on the blank (called rough features) with a minimum machining allowance which ensures the required surface finish.
+The allowance calculation requires the final part specification in the form of the machining CNC code as well as a representation of the rough geometries.
+From their positions and geometric parameters their distance can be computed, which then can be used to generate the allowance constraints for the optimization program.
+
[…]
