+++ paper/ref.bib
@@ -86,3 +86,11 @@
   urldate = {2023-10-09},
   copyright = {All rights reserved}
 }
+
+@book{Hoffimann2023,
+  title = {Geospatial Data Science with {{Julia}}},
+  author = {Hoffimann, Júlio},
+  year = {2023},
+  doi = {10.5281/zenodo.10150870},
+  url = {https://juliaearth.github.io/geospatial-data-science-with-julia}
+}
