+++ paper/header.tex
@@ -4,10 +4,11 @@
 
 \author[1, 2]{Tamás Cserteg}
 \author[1]{András Kovács}
-\author[1, 3]{József Váncza}
+\author[1]{József Váncza}
+%\author[1, 3]{József Váncza}
 \affil[1]{HUN-REN Institute for Computer Science and Control (SZTAKI), Budapest, Hungary}
 \affil[2]{Doctoral School of Informatics, ELTE Eötvös Loránd University, Budapest, Hungary}
-\affil[3]{Department of Manufacturing Science and Technology, Budapest University of Technology and Economics, Budapest, Hungary}
+%\affil[3]{Department of Manufacturing Science and Technology, Budapest University of Technology and Economics, Budapest, Hungary}
 
 \keywords{Julia, Optimization, Machining, Blank localization}
 
