+++ paper/paper.yml
@@ -13,13 +13,11 @@ authors:
     affiliation: 1
   - name: József Váncza
     orcid: 0000-0003-0054-0560
-    affiliation: 1, 3
+    affiliation: 1
 affiliations:
   - name: HUN-REN Institute for Computer Science and Control (SZTAKI), Budapest, Hungary
     index: 1
   - name: Doctoral School of Informatics, ELTE Eötvös Loránd University, Budapest, Hungary
     index: 2
-  - name: Department of Manufacturing Science and Technology, Budapest University of Technology and Economics, Budapest, Hungary
-    index: 3
-date: 24 May 2024
+date: 10 April 2025
 bibliography: ref.bib
