+++ paper/ref.bib
@@ -12,16 +12,6 @@
   urldate = {2022-04-27}
 }
 
-@inproceedings{ding:2021_CoarsefineOptimizationMethod,
-  title = {A {{Coarse-fine Optimization Method}} for {{Blank Localization}} of {{Complex Surfaces}} Based on {{3D Point Clouds}}},
-  booktitle = {2021 {{IEEE}} 2nd {{International Conference}} on {{Information Technology}}, {{Big Data}} and {{Artificial Intelligence}} ({{ICIBA}})},
-  author = {Ding, Ji and Kang, Xubing and Zhou, Shuaijun and Jia, Yujiao and Tan, Yaxiong and Yan, Xinying},
-  year = {2021},
-  volume = {2},
-  pages = {1099--1105},
-  doi = {10.1109/ICIBA52610.2021.9688202}
-}
-
 @article{cserteg:2023_Annals,
   title = {Multi-Operation Optimal Blank Localization for near Net Shape Machining},
   author = {Cserteg, Tam{\'a}s and Kov{\'a}cs, Andr{\'a}s and V{\'a}ncza, J{\'o}zsef},
