### 2025-04-10 a7dbedf
remove comments and old text chunks from paper
