+++ paper/paper.tex
@@ -6,7 +6,6 @@
 \begin{document}
 
 \input{header}
-%TODO: ha kell hely, akkor Jóska BME-t törölhetjük
 
 \maketitle
 
@@ -15,7 +14,6 @@
 Blank localization (also known as workpiece referencing) is an essential task in machining.
 It aims to precisely establish the geometric relation of the machine tool (mill, lathe, etc.) and the workpiece.
 We introduced the concept of multi-operation blank localization to address this task for drilling and milling scenarios in a semi-automated way,
-%TODO: innentől a mondat második fele cut?
 which allows positioning different machining features (e.g., different holes) separately in order to exploit the tolerances on the relative position of those features to compensate the small errors of the blank.
 The method takes as input the measured rough geometry and the machining CNC code, and computes the best possible position of each feature considering machining allowances and tolerances by solving a convex quadratically constrained quadratic program (QCQP).
 The versatility and extensibility of the Julia language helped the development of this algorithm, materializing in the \texttt{BlankLocalizationCore.jl} package.
@@ -26,8 +24,6 @@ Its flexibility and ease of use make it an excellent research tool that can be d
 \label{sec:intro}
 Cast parts may have small geometric variations from lot to lot that need to be addressed before machining by altering the CNC code.
 Current practice is dominated by iterative adjustments by human operators, which requires highly trained workers and takes a long time.
-%, and still, may produce scrap.
-%TODO: kell egyáltalán ezeket hivatkozni?
 Automated methods exist for complex free-form parts like wind turbines that place the entire blank as a single solid object \cite{tan:2014_UnconstrainedApproachBlank}.
 Multi-operation blank localization~\cite{cserteg:2023_Annals} however handles groups of features independently, providing greater flexibility than traditional approaches.
 It focuses on drilling and milling which are among the most common machining operations, making it applicable to a wide range of products.
@@ -38,27 +34,11 @@ Exactly for these reasons we chose the Julia language~\cite{bezanson2017julia}.
 \label{sec:algo}
 
 The problem involves looking for the optimal position for each to-be-machined feature (short: machined feature) on the workpiece.
-%chatgpt
 These features are grouped together, with each group defined relative to a specific reference point known as the part zero.
 The layout of these part zeros is controlled by the structure of the machining CNC code.
 By moving the part zeros, we can indirectly control the positions of the associated features.
 Each part zero—and by extension, each group—can be moved independently, which gives the method its flexibility.
 In the optimization program, the positions of the part zeros are the decision variables.
-
-%eredeti
-\iffalse
-These features are structured into groups, each group being defined relative to a common reference, called part zero.
-This is defined by the structure of the machining CNC code.
-This means that we can influence the features' positions through moving the part zeros.
-Part zeros (thus the groups) can be moved separately, which gives the method its flexibility.
-The position of the part zeros are the decision variables of the optimization program.
-\fi
-%Features within the same group must be moved together, but different feature groups can be moved separately.
-%TODO: optimalizáló leírás:
-% 0. motiváció
-% 1. mik a döntések
-% 2. mik a korlátok
-% 3. mi a célfv
 
 To ensure the required surface finish, the machined features must enclose the pre-cast features on the blank (called rough features) with a minimum machining allowance (a lower bound parameter).
 The allowance calculation requires the final part specification in the form of the machining CNC code as well as a representation of the rough geometries.
@@ -66,104 +46,35 @@ From their positions and geometric parameters their distance can be computed, wh
 
 The other set of constraints roots from respecting the dimensional tolerances describing functional properties (e.g. connections to mating parts).
[…]
