### 2025-04-10 703b161
update future plans to match article updates
