+++ paper/paper.tex
@@ -127,9 +127,9 @@ end
 \section{Results and future work}
 \label{sec:results}
 
-Future plans for the package and the method itself include an overhaul of the tolerance modeling scheme.
-Currently, only dimensional tolerances are handled, but it should be discovered if more GD\&T tolerances can be incorporated into the model.
-The current implementation supports only cylinders and planes, thus drilling and milling; it is a future research direction to extend the handled machining operations, e.g., to turning.
+Future plans for the package and the method itself include an overhaul of the tolerance modelling scheme.
+It should be possible to generalize the current dimensional tolerances for  more GD\&T tolerances by introducing a new type hierarchy based on the current geometry type tree and a function interface for the optimization program.
+Another direction of development is to extend the supported operations from drilling and milling to others, such as turning, although this is more a research question than an implementation.
 
 \section{Acknowledgments}
 The research was supported by the European Union within the framework of the National Laboratory for Autonomous Systems (RRF-2.3.1-21-2022-00002) and the TKP2021-NKTA-01  NRDIO grant on "Research on cooperative production and logistics systems to support a competitive and sustainable economy".
