+++ previews/PR39/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#Using-the-API-to-extend-the-geometry-types","page":"API extension","title":"Using the API to extend the geometry types","text":"","category":"section"},{"location":"api/","page":"API extension","title":"API extension","text":"The solver handles hole and face (plane) features that are either IsPrimitive or IsFreeForm. A few basic types are defined like SimpleHole and MeshHole, but new ones can be also defined. For example for point clouds, that don't have faces, like a mesh, only points.","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"This capability is properly documented, as a major API rewrite is going on in #3.","category":"page"},{"location":"api/#Defining-a-new-geometry","page":"API extension","title":"Defining a new geometry","text":"","category":"section"},{"location":"api/","page":"API extension","title":"API extension","text":"Defining a new geometry is pretty easy:","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"It has to be the subtype of either AbstractHoleGeometry or AbstractPlaneGeometry.\nThe GeometryStyle trait is need to be defined: it is either a IsPrimitive or IsFreeForm geometry.\nIsFreeForm geometries need to define the surfacepoints and filteredsurfacepoints functions.\nIsPrimitive geometries need to define the featurepoint functions\nThose IsPrimitive geometries that are subtype of AbstractHoleGeometry also need to define the featureradius function\nFinally, for visualization purposes the visualizationgeometry function should be defined, that returns an object, that can be used with the Meshviz.viz() function.","category":"page"},{"location":"api/#An-example","page":"API extension","title":"An example","text":"","category":"section"},{"location":"api/","page":"API extension","title":"API extension","text":"Let's assume that we want to define a new holelike type: MyDisk, which has a featurepoint, a normal, and a radius.","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"using BlankLocalizationCore\n\nstruct MyDisk <: AbstractHoleGeometry\n    point::Vector{Float64}\n    normal::Vector{Float64}\n    diameter::Float64\nend\n\nGeometryStyle(::Type{MyDisk}) = IsPrimitive()","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"It is an IsPrimitive and holelike feature, therefore we need to define the:","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"featurepoint\nfeatureradius\nvisualizationgeometry","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"functions. These look like as follows:","category":"page"},{"location":"api/","page":"API extension","title":"API extension","text":"featurepoint(::IsPrimitive, x::MyDisk) = x.point\nfeatureradius(::IsPrimitive, x::MyDisk) = x.diameter/2\n\nusing Meshes\n\nfunction visualizationgeometry(geom::MyDisk)\n    plane = Plane(Point3(geom.point), Vec3(geom.normal))\n    return Disk(plane, geom.diameter/2)\nend\n","category":"page"},{"location":"reference/#Reference-for-exported-types-and-functions","page":"Reference","title":"Reference for exported types and functions","text":"","category":"section"},{"location":"reference/#API-docstrings","page":"Reference","title":"
[…]
