### 2025-04-11 #182 @cserteGT3
Thank you once again, @mzagorowska, for taking the time to review our work. Your valuable comments and questions have guided our revisions and, hopefully, made the paper easier to understand. The [last build above](https://github.com/JuliaCon/proceedings-review/issues/182#issuecomment-2796617113) by editorialbot is the latest version. Please find our responses to your comments below.

>
>     * (installation, installation instructions) I had to add packages (ipopt, GLMakie, Meshes) from errors when trying to run the example

Sorry for the inconvinience. We've updated the installation instructions and the example code. As it is in a separate branch for now, this updated documentation is only avaliable via [this preview build](https://csertegt3.github.io/BlankLocalizationCore.jl/previews/PR39/).

>     * (performance) It is not entirely clear to me whether the improvements shown in Fig 2 are from the algorithm published already in Ref. [4], or from the implementation in Julia in the current work. Also, the plots in the current paper seem to be taken 1:1 from Ref [4]
>     * (methodology) The approach is justified, the trouble is with reproducibility based purely on the paper.

We've removed that figure from the paper and now we don't make any performance claims. (Just for the record, the mentioned improvements came from the method and not from the Julia implementation. I understand that this was not clearly written.)

>     1. The paper refers to Ref [4] for the algorithm and I would find it difficult to say what exactly is done without Ref [4].
>
>     2. I would also like to see a bit more of detail on the interplay between the packages and the custom extensions that made using diverse packages, such as JuMP with Meshes, possible.
>
>     3. It could also be a good addition to say how the elements from Section 2 were implemented in Julia (at least mention the relevant functions). Right now, the first part is a rather theoretical introduction to the problem, without indicating the solutions in Julia.

Section 2 has been updated to provide a more concise description of the method and the optimization problem. Section 3 now focuses on the implementation details and the interplay of the mentioned packages.

>     4. Regarding future plans, it would also be interesting to see how the proposed extensions can be included in the package, if they need substantial changes.

Section 4 has also been updated to include some detail on how we would approach the proposed extensions.

Thank you again for the review and please let me know if you have any questions or comments.
