### 2025-05-05 cserteGT3/BlankLocalizationCore.jl#40 @Nimrais
Review Juliacon Proceedings
# Review of *BlankLocalizationCore.jl*

Hi, you have a really nice demonstration of how Julia can be applied to a real-world machining workflow. I’m delighted to see more domain-specific packages being created and maintained in the Julia community. Below is my feedback, with each item marked as **Critical** (must-fix before acceptance) or **Recommended** (would strongly improve the work but is not a hard blocker).

---

## A  Paper / Documentation

| # | Critical ? | Observation | Recommendation |
|:--:|:----------:|-------------|----------------|
| **A1** | Recommended | The text assumes readers know both Julia idioms and machining jargon. | Add a *“Quick-start for non-Julia machining engineers”* sidebar that lists the few Julia constructs they need to understand to model the domain in Julia (`struct`, `optimize!`, …) and shows one-line installation:<br>`import Pkg; Pkg.add("BlankLocalizationCore")`. |
| **A2** | ✔︎ | No minimal working example (MWE) appears in the README or paper; new users can’t see how to build a model from scratch. | Embed a **15-line 2-D example** (shift-only toy) in the README. |
| **A3** | Recommended | Figures show full 3-D aerospace parts, which can intimidate newcomers. | Prepend a *cartoon 2-D diagram* (like the square-and-shift plot) so the core idea is clear before industrial complexity. |
| **A4** | Recommended | Terminology leans on optimisation references that machining readers may not know. | Add a **reverse glossary** (e.g. *cone → radial clearance check*). |
---

## B  Code / Repository & CI

| # | Critical ? | Observation | Recommendation |
|:--:|:----------:|-------------|----------------|
| **B1** | ✔︎ | **No CI pipeline**. | Enable **GitHub Actions** with the standard Julia CI template, testing on `1.11` and `nightly`. |
| **B3** | ✔︎ | CI Badge is not working. | You should fix the CI Badge.|
| **B4** | Recommended | Package installs heavy solvers by default. | You can use julia extentsion to load specific code based on this solvers only if the solver's Pacakge is presented in the current environment. that https://pkgdocs.julialang.org/v1/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions) |

---

### Summary

* **Critical before acceptance**: add a minimal 2-D example and beginner sidebar (A2); set up CI and a basic test suite (B1, B2).
* **Recommended**: gentler visuals, glossary, limitations section, documentation badge, solver configurability, repo slimming.
