### 2025-05-05 6dbf7d5
use CI code generated by PkgTemplates
