### 2025-05-05 f91ee3c
fix current lts: 1.6->1.10
