+++ .github/workflows/CI.yml
@@ -23,8 +23,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
+          - '1.10'
           - '1.11'
-          - '1.6'
           - 'pre'
         os:
           - ubuntu-latest
