+++ README.md
@@ -38,6 +38,6 @@ Please open an issue if you encounter any problems. We take issues seriously and
 
 This package couldn't have been created without the great people behind the following projects (as well as the whole Julia ecosystem):
 
-* [Meshes.jl](https://github.com/JuliaGeometry/Meshes.jl) and the [Makie.jl](https://github.com/MakieOrg/Makie.jl) ecosystem
-* [JuMP.jl](https://jump.dev/)
-* [DataFrames.jl](https://github.com/JuliaData/DataFrames.jl) and [PrettyTables.jl](https://github.com/ronisbr/PrettyTables.jl)
+- [Meshes.jl](https://github.com/JuliaGeometry/Meshes.jl) and the [Makie.jl](https://github.com/MakieOrg/Makie.jl) ecosystem
+- [JuMP.jl](https://jump.dev/)
+- [DataFrames.jl](https://github.com/JuliaData/DataFrames.jl) and [PrettyTables.jl](https://github.com/ronisbr/PrettyTables.jl)
