### 2025-05-05 2548230
run CI on nightly instead of prereleases
