+++ .github/workflows/CI.yml
@@ -25,7 +25,7 @@ jobs:
         version:
           - '1.10'
           - '1.11'
-          - 'pre'
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
