### 2025-05-05 30da6be
remove meshviz from docs as well
