+++ README.md
@@ -41,7 +41,3 @@ This package couldn't have been created without the great people behind the foll
 * [Meshes.jl](https://github.com/JuliaGeometry/Meshes.jl) and the [Makie.jl](https://github.com/MakieOrg/Makie.jl) ecosystem
 * [JuMP.jl](https://jump.dev/)
 * [DataFrames.jl](https://github.com/JuliaData/DataFrames.jl) and [PrettyTables.jl](https://github.com/ronisbr/PrettyTables.jl)
-
-<!---
-To build the paper, use this site: https://whedon.theoj.org/
---->
