### 2025-05-13 74cad80
add missing docstrings for Tolerance
