### 2025-05-13 e2602f5
add shorter MWE to Readme and documentation
