+++ docs/src/example-2d.md
@@ -1,0 +1,187 @@
+# Short 2D example
+
+A shorter 2D example is introduced in this section.
+It provides the basics for easier understanding of the [Complex 3D example](@ref).
+
+## Part description
+
+The task is to find optimal positions for drilling the three holes on one side of the part.
+The three holes are machined in one operation: the machining poses are defined to a common part zero.
+The distance of the holes and another side of the part is toleranced (only along one, the x axis).
+In this example we are searching the optimal x value of the part zero (and are ignoring the y-z values for simplicity).
+We mean optimal in the sense that, 1) material must be removed at every hole (the blue circles must cover the black ones with enough margin called machining allowance) and 2) the toleranced distances between the left plane and the holes should be as close to the center of the tolerance fields as possible.
+Please note, that the package is designed for 3D usage, therefore the geometry defintions need to have all coordinates defined (even though this example is a planar problem).
+
+![example-2d](assets/example-2d.png)
+
+## Package installations
+
+The BlankLocalizationCore package and one solver (Ipopt, Xpress, etc.) needs to be installed to use the package.
+
+```julia
+using Pkg
+Pkg.add(["BlankLocalizationCore", "Ipopt"])
+```
+
+## Model construction
+
+After `using` the package, we can move on the model building:
+
+```julia
+using BlankLocalizationCore
+```
+
+Please note, that the package uses 3D geometry 
+
+First, the part zero is created with the following line:
+
+```julia
+# partzero relative to origin:
+# translation: zero, will be calculated by optimization
+# orientation: given as rotation matrices
+pz_holes = PartZero("pz-holes", [0,0,0], hcat([1,0,0], [0,1,0], [0,0,1]))
+```
+
+### Machined geometries
+
+The CNC code for the holes is modeled with cylinders.
+A [`SimpleHole`](@ref) struct stores the position of a cylinder relative to the part zero and its radius.
+
+```julia
+# machined features: 3 holes + 1 plane
+h1 = SimpleHole([62, 7, 0], 11.5)
+h2 = SimpleHole([80, 52, 0], 13)
+h3 = SimpleHole([108, 26, 0], 12.5)
+p1 = SimplePlane([0, 0, 0]) # this is just a dummy machined plane for the left side
+```
+
+The left side of the part is not machined in this example, but a dummy variable is needed for the modeling, hence the above `SimplePlane` definition.
+
+### Rough geometries
+
+Similarly, the to-be-machined (rough) geometries are also defined as primitive geometries.
+Their position is relative to the origin.
+
+```julia
+# rough features: 3 holes and left side of the part are measured geometries for the model
+# measured relative to origin
+rh1 = HoleAndNormal([-56.79, 44.235, 0], [0, 0, -1], 11.2)
+rh2 = HoleAndNormal([-38.79, 89.235, 0], [0, 0, -1], 12.8)
+rh3 = HoleAndNormal([-10.79, 63.235, 0], [0, 0, -1], 12.2)
+rp1 = PlaneAndNormal([-125.67, 66, 0], [-1, 0, 0])
+```
+
+The machined and rough geometries are connected via [`FeatureDescriptor`](@ref)s, that defines the name of the feature, and also if it is machined and if it has a rough geometry (for cases when holes don't have cast holes on the blank).
+
+Then, from the rough, machined geometries and feature descriptors the [`LocalizationFeature`](@ref)s are constructed.
+
+```julia
+# feature descriptors
[…]
