### 2025-05-13 6a65515
add shorter MWE to Readme and documentation
