+++ previews/PR39/siteinfo.js
@@ -1,1 +1,2 @@
 var DOCUMENTER_CURRENT_VERSION = "previews/PR39";
+var DOCUMENTER_IS_DEV_VERSION = false;
