### 2025-05-13 765003e
add missing code lines for 2D example
