### 2025-05-13 cserteGT3/BlankLocalizationCore.jl#40 @cserteGT3
I think I'm sufficiently satisfied with the current state. Thank you again for the detailed and structured review. I'm answering the raised issues below:

* A1 & A4: great suggestions! I've opened an issue to tackle is later as currently I am more focused towards finishing my research
* A2: I've created a new, easier example. It is still not a 15-line example, but a much simpler problem and illustration. It is included in the documentation ([preview here](https://csertegt3.github.io/BlankLocalizationCore.jl/previews/PR39/example-2d/) and in a bit shorter form in the readme as well.
* A3: are you thinking of the Example part section (here: https://csertegt3.github.io/BlankLocalizationCore.jl/v2.3/example/#Example)? If so, this is a great point. I'm not good at visualizing the concept, but I'll ask some colleagues if they could help me out.
* B1: I've updated the CI code in this branch, so the code uses a newer versions of `julia-actions`. The code is tested on 1.10 (lts), 1.11 and nightly
* B2: the last build happened to be too old for the badge to be green. I fixed it with triggering a build on main, but the fixes for B1 in general help with CI maintenance.
* B3: I don't see you point here. Solvers are not loaded in the package code and JuMP is a necessary dependency for implementing the optimization program. Therefore I would not want to move JuMP to package extension. The use of DataFrames (the other big dependency) however should be reconsidered as it is only needed for the result evaluation. Again, I don't want to move it to a package extension, that feels weird, but removing it would be good idea. I've opened an issue to track this: #41

If you have any comments on my answers and solutions, please let me know!
