### 2025-05-13 #182 @cserteGT3
Thank you @Nimrais for the review once more! As the questions were raised in a separate [issue](https://github.com/cserteGT3/BlankLocalizationCore.jl/issues/40), I answered them over there. Though I copy here my answers to the 3 blocking requests:
* A2: I've created a new, easier example. It is still not a 15-line example, but a much simpler problem and illustration. It is included in the documentation ([preview here](https://csertegt3.github.io/BlankLocalizationCore.jl/previews/PR39/example-2d/) and in a bit shorter form in the readme as well (still everything in the `paper` branch).
* B1: I've updated the CI code in the `paper` branch, so the code uses a newer versions of `julia-actions`. The code is tested on 1.10 (lts), 1.11 and nightly.
* B2: the last build happened to be too old for the badge to be green. I fixed it with triggering a build on main, but the fixes for B1 in general help with CI maintenance.

Let me know if you accept my answers for these in particular and my edits in general.
