### 2025-05-13 cserteGT3/BlankLocalizationCore.jl#40 @Nimrais
HI!

> A1 & A4: great suggestions! I've opened an issue to tackle is later as currently I am more focused towards finishing my research

Sure no big deal.

> A2: I've created a new, easier example. It is still not a 15-line example, but a much simpler problem and illustration. It is included in the documentation ([preview here](https://csertegt3.github.io/BlankLocalizationCore.jl/previews/PR39/example-2d/) and in a bit shorter form in the readme as well (still everything in the paper branch).

I checked it, it is far better. I am considering this issue resolved.

> A3: are you thinking of the Example part section (here: https://csertegt3.github.io/BlankLocalizationCore.jl/v2.3/example/#Example)? If so, this is a great point. I'm not good at visualizing the concept, but I'll ask some colleagues if they could help me out.

The new 2D picture is quite nice already.

> B1: I've updated the CI code in the paper branch, so the code uses a newer versions of julia-actions. The code is tested on 1.10 (lts), 1.11 and nightly

Good!

> B2: the last build happened to be too old for the badge to be green. I fixed it with triggering a build on main, but the fixes for B1 in general help with CI maintenance.

Yeah I see.

> B3: I don't see you point here. Solvers are not loaded in the package code and JuMP is a necessary dependency for implementing the optimization program. Therefore I would not want to move JuMP to package extension. The use of DataFrames (the other big dependency) however should be reconsidered as it is only needed for the result evaluation. Again, I don't want to move it to a package extension, that feels weird, but removing it would be good idea. I've opened an issue to track this: https://github.com/cserteGT3/BlankLocalizationCore.jl/issues/41

It's a recommendation, so there's no need to implement it right now if you don't want to. Also, I think now you are right, there is no direct benefit to abstracting JuMP right away. It would probably be hard and there wouldn't be much utility out of this abstraction, I agree.

DataFrames, in my taste, is a really heavy dependency, and you are using it only for the result evaluation which is not the direct offering of your repo, so it's an ideal candidate to be an extension.

But generally I am considering my review addressed, good work.
