### 2025-05-14 #182 @mzagorowska
All good on my side. A minor note - the documentation in the branch "paper" states two ways of installing the package, one with Ipopt and one without. The one without is shown first but is insufficient to run the example (complains about missing ipopt). The second one works.
