### 2025-05-14 f8f268b
update installation steps and guide links in readme
