### 2025-05-14 7abf737
update installation instructions in documentation to follow readme
