+++ docs/src/index.md
@@ -29,15 +29,28 @@ When doing so, one must consider two important factors:
 
 Our paper proposes a method, that ensures a proper machining allowance (minimum requirement), while trying to optimize to the center of the tolerance fields between features.
 
-The documentation goes through a detailed example of the process while showing how to use the package.
+## Installation and usage
 
 The package is registered in the general registry, so it can be installed via running:
 
 ```julia
-] add BlankLocalizationCore
+using Pkg
+Pkg.add("BlankLocalizationCore")
 ```
 
-For the exaplanation on how the package works, please read through the [Short 2D example](@ref) and [Complex 3D example](@ref) pages.
+Note, that an optimization solver is needed to use the package (Ipopt or Xpress for example), which needs to be installed as well:
+
+```julia
+Pkg.add("Ipopt") # open source solver
+# Pkg.add("Xpress") # commercial solver, that requires purchased/community license
+```
+
+For the exaplanation on how the package works, see the followings:
+
+- [First steps](https://github.com/cserteGT3/BlankLocalizationCore.jl?tab=readme-ov-file#first-steps) in the Readme: short example with copy-paste code.
+- [Short 2D example](@ref): the same example in the documentation with a bit more details.
+- [Complex 3D example](@ref): a complex, detailed 3D example showcasing the full potential of the package.
+
 
 ## Acknowledgements
 
