### 2025-05-14 #182 @cserteGT3
> A minor note - the documentation in the branch "paper" states two ways of installing the package, one with Ipopt and one without. The one without is shown first but is insufficient to run the example (complains about missing ipopt). The second one works.

Thank you for pointing out! I updated the instructions (in the readme and the documentation page too) to better describe the solver requirement. And added a oneliner for installation.

Once again, thank you @mzagorowska and @Nimrais for the reviews! I believe that with your comments and requests the paper has been greatly improved.
