### 2025-05-22 36fb69b
release v2.4.1 to trigger zenodo
