+++ Project.toml
@@ -1,7 +1,7 @@
 name = "BlankLocalizationCore"
 uuid = "dc39d501-6ffa-4c27-9708-7dc7e3b1a5c3"
 authors = ["Tamás Cserteg <cserteg.tamas@sztaki.hu>", "András Kovács <akovacs@sztaki.hu> and contributors"]
-version = "2.4.0"
+version = "2.4.1"
 
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
