### 2025-05-22 #182 @cserteGT3
I've:
- picked the commits that modified the repository (except the `paper`) folder onto the main branch
- created a new release: [v2.4.1](https://github.com/cserteGT3/BlankLocalizationCore.jl/releases/tag/v2.4.1)
- created a zenodo archive from the release: https://zenodo.org/records/15489856, DOI: `10.5281/zenodo.15489856`
